refactor(import): move import parsing out of the settings UI module
About a thousand lines of GlassEQSettingsUI never touched SwiftUI: bounded
text file reads, the stereo pair merger, name validation, the format sniffer,
the WAV decoder, and the AutoEq catalogue client. They now live in
GlassEQCore and a new GlassEQProfileImport target, so the IPC test target no
longer links the UI module to test a parser. The import sheet itself is
split into the sheet, two panes, the source summaries, an import model that
owns the single commit path, and the open-panel wrapper.

Along the way: the AutoEq results list filters once per query instead of
twice per body pass, the catalogue reload uses one view-scoped task instead
of racing a stored one, the panes dismiss through the environment, the
download delegate is replaced by a byte-capped session stream, and both
sample-rate labels share one locale-aware formatter.
